requirements.txt
matplotlib-inline==0.1.3
numpy==1.21.3
pandas==1.3.4
streamlit==1.1.0
streamlit-drawable-canvas==0.8.0
tensorflow==2.7.0